// src/app/events/enums/events.enum.ts
export enum EventsEnum {
	TOKEN_VERIFIED = "TOKEN_VERIFIED",
	TOKEN_BOUGHT = "TOKEN_BOUGHT",
	TOKEN_SELLED = "TOKEN_SELLED",
	TOKEN_START_BUYING = "TOKEN_START_BUYING",
	TOKEN_START_SELLING = "TOKEN_START_SELLING",
	SIGNAL_CREATED = "SIGNAL_CREATED",
	SIGNALS_CREATED = "SIGNALS_CREATED",
	TOKEN_CREATED = "TOKEN_CREATED",
	STRATEGY_CREATED = "STRATEGY_CREATED",
	STRATEGY_UPDATED = "STRATEGY_UPDATED",
	STRATEGIES_CREATED = "STRATEGIES_CREATED",
	CONDITIONS_GROUPS_CREATED = "CONDITIONS_GROUPS_CREATED",
	CONDITIONS_GROUP_CREATED = "CONDITIONS_GROUPS_CREATED",
	CONDITIONS_GROUP_UPDATED = "CONDITIONS_GROUP_UPDATED",
	MILESTONE_CREATED = "MILESTONE_CREATED",
	MILESTONES_CREATED = "MILESTONES_CREATED",
	MILESTONE_UPDATED = "MILESTONE_UPDATED",
	CONDITIONS_CREATED = "CONDITIONS_CREATED",
	CONDITION_CREATED = "CONDITION_CREATED",
	CONDITION_UPDATED = "CONDITION_UPDATED",
	TRADING_TOKEN_CREATED = "TRADING_TOKEN_CREATED",
	TRADING_TOKENS_CREATED = "TRADING_TOKENS_CREATED",
	TRADING_TOKEN_UPDATED = "TRADING_TOKEN_UPDATED",
	TRADING_TOKEN_DELETED = "TRADING_TOKEN_DELETED",
	HELIUS_MESSAGE = "HELIUS_MESSAGE",
	HELIUS_OPEN = "HELIUS_OPEN",
	HELIUS_CLOSE = "HELIUS_CLOSE",
	HELIUS_ERROR = "HELIUS_ERROR",
	SOLANA_TRANSACTION = "SOLANA_TRANSACTION",
	TRANSACTION_CREATED = "TRANSACTION_CREATED",
	TRANSACTION_UPDATED = "TRANSACTION_UPDATED",
	TRANSACTION_DELETED = "TRANSACTION_DELETED",
	TRANSACTIONS_CREATED = "TRANSACTIONS_CREATED",
	CANDLE_CREATED = "CANDLE_CREATED",
	CANDLES_CREATED = "CANDLES_CREATED",
	CANDLE_UPDATED = "CANDLE_UPDATED",
	CANDLE_DELETED = "CANDLE_DELETED"
}